+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [LIBRARY] add a original line number to rulers.
 - [LIBRARY] correct line numbering on nested list items.
-- [LIBRARY] correct line numbering of block-level elements inside paragraphs and table's cells.
+- [LIBRARY] correct line numbering of block-level elements inside paragraphs, table's cells, definition lists.
 
 
 ## [2.0.105] - 20200314